requirement.txt
Django
django-colorfield
django-cors-headers
django-grappelli
django-mptt
django-rest-auth
djangorestframework
psycopg2
pytz
PyYAML
six
